+++ tests/testthat/test-vr_vec.R
@@ -1,8 +1,6 @@
 
 test_that("vr_vec functions work correctly", {
   
-  library(popbio)
-  
   ### test vr_vec_U
   U1 <- rbind(c(  0,   0,   0,   0),
               c(0.5, 0.2, 0.1,   0),
@@ -65,7 +63,7 @@ test_that("vr_vec functions work correctly", {
   expect_true(all(is.na(fecu1[1:2])))
   
   # weighting
-  v <- reproductive.value(U1 + R1)
+  v <- popdemo::eigs(U1 + R1, what = "rv") #reproductive value
   fecu2 <- vr_vec_reproduction(U1, R1, weights_row = v)
   expect_true(fecu2[3] != fecu1[3])