+++ NEWS.md
@@ -3,6 +3,7 @@
 * Improved documentation of `mpm_to_table` and related functions.
 * Standardised the `lx_crit`, `conv` and `xmax` argument values across age-from-stage functions.
 * Minor edits to documentation for stylistic consistency.
+* Removed dependency on the package `popbio`.
 
 
 # Rage 1.1.0